+++ src/main/java/org/octri/omop_annotator/repository/omop/NoteRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(path = "note")
-public interface NoteRepository extends PagingAndSortingRepository<Note, Long> {
+public interface NoteRepository extends PagingAndSortingRepository<Note, Integer> {
 
 	static final String query = "select note.id as id, note.person.id as person,"
 			+ " noteTypeConcept.name as type, note.noteDatetime as date, noteClassConcept.name as noteClass,"
@@ -20,9 +20,9 @@ public interface NoteRepository extends PagingAndSortingRepository<Note, Long> {
 			+ " left join note.visitOccurrence visitOccurrence";
 
 	@Query(query + " where note.person.id = ?1")
-	List<NoteRow> findByPersonId(Long id);
+	List<NoteRow> findByPersonId(Integer id);
 
 	@Query(query + " where visitOccurrence.id = ?1")
-	List<NoteRow> findByVisitOccurrenceId(Long visitOccurrenceId);
+	List<NoteRow> findByVisitOccurrenceId(Integer visitOccurrenceId);
 
 }
